requirements.txt
# FloodBrief — pinned dependencies
# Tested on Python 3.11, CUDA 12.1, Ubuntu 22.04 / Colab T4

# === TerraMind + TerraTorch ===
terratorch>=1.2.4
diffusers==0.30.0

# === Core PyTorch ===
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0

# === HuggingFace ===
transformers>=4.40.0
huggingface-hub>=0.22.0
datasets>=2.16.0

# === Geospatial ===
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.6.0
geopandas>=0.14.0

# === Scientific Python ===
numpy>=2.2.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# === Plotting ===
matplotlib>=3.7.0

# === Demo ===
gradio>=4.0.0

# === Utilities ===
tqdm>=4.65.0
Pillow>=10.0.0
pyyaml>=6.0.0
gdown>=5.0.0
